+++ src/App.tsx
@@ -3,7 +3,7 @@ import { Nav } from 'components/common/Nav'
 import { Home } from 'components/pages/Home'
 import { Guide } from 'components/pages/Guide'
 import { Module } from 'components/pages/Module'
-import { useModules } from 'hooks/useModules'
+import { useModules } from 'hooks/module/useModules'
 import { ConfigProvider, Layout } from 'antd'
 import { Content } from 'antd/es/layout/layout'
 import { urlify } from 'utils/urlify'
