+++ src/hooks/module/useAddModule.tsx
@@ -1,29 +1,29 @@
 import { useMutation, useQueryClient } from 'react-query'
 import axios from 'axios'
 import { type IconType } from 'antd/es/notification/interface'
-import { type CourseModuleCreateEditModel } from 'schemas/PhraseCRUD'
+import { type CourseModuleCreateEditModel } from 'schemas/ModuleCRUD'
 
 const apiUrl =
   process.env.REACT_APP_API_URL ??
   'https://persian-paradise-api-production.up.railway.app'
 
-const updateModule = async (module: CourseModuleCreateEditModel) => {
-  const response = await axios.put(`${apiUrl}/api/module/update`, module)
+const addModule = async (module: CourseModuleCreateEditModel) => {
+  const response = await axios.post(`${apiUrl}/api/module/add`, module)
   return response.data
 }
 
-export const useUpdateModule = (
+export const useAddModule = (
   openNotification: (message: string, type: IconType) => void
 ) => {
   const queryClient = useQueryClient()
 
-  return useMutation(updateModule, {
+  return useMutation(addModule, {
     onSuccess: async () => {
-      openNotification('Successfully Updated Module', 'success')
+      openNotification('Successfully Added Module', 'success')
       await queryClient.invalidateQueries('modules')
     },
     onError: async () => {
-      openNotification('Could Not Update Module', 'error')
+      openNotification('Could Not Add Module', 'error')
       await queryClient.invalidateQueries('modules')
     }
   })
